Rename handlers js to ts
