+++ packages/multichain/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@metamask/api-specs": "^0.10.12",
     "@metamask/eth-json-rpc-filters": "^7.0.0",
+    "@metamask/json-rpc-engine": "^9.0.3",
     "@metamask/rpc-errors": "^6.3.1",
     "@open-rpc/schema-utils-js": "^2.0.5",
     "lodash": "^4.17.21"
