+++ packages/multichain/src/adapters/caip-permission-adapter-middleware.ts
