+++ packages/multichain/src/handlers/wallet-revokeSession.ts
@@ -1,16 +1,21 @@
+import type { JsonRpcEngineNextCallback, JsonRpcEngineEndCallback } from '@metamask/json-rpc-engine';
 import {
   PermissionDoesNotExistError,
   UnrecognizedSubjectError,
+  PermissionController
 } from '@metamask/permission-controller';
 import { rpcErrors } from '@metamask/rpc-errors';
-import { Caip25EndowmentPermissionName } from '../caip25permissions';
+import { Caip25EndowmentPermissionName } from '../caip25Permission';
+import { JsonRpcRequest, JsonRpcResponse } from '@metamask/utils';
 
 export async function walletRevokeSessionHandler(
-  request,
-  response,
-  _next,
-  end,
-  hooks,
+  request: JsonRpcRequest<Params>,
+  response: JsonRpcResponse<true>,
+  _next: JsonRpcEngineNextCallback,
+  end: JsonRpcEngineEndCallback,
+  hooks: {
+    revokePermission: PermissionController['revokePermission']
+  },
 ) {
   try {
     hooks.revokePermission(request.origin, Caip25EndowmentPermissionName);
